+++ gala/potential/potential/__init__.py
@@ -5,3 +5,17 @@ from .builtin import *
 from .io import *
 from .util import *
 from .interop import *
+
+
+def __getattr__(name):
+    # Needed for MultipolePotential save/load
+    from . import builtin
+
+    if name in globals():
+        return globals()[name]
+
+    elif name.startswith('MultipolePotentialLmax'):
+        return getattr(builtin.core, name)
+
+    else:
+        raise AttributeError("huh")